To prevent excessive code duplication (or severe brain injury) I split up the gossiping proxies and the application data routing/forwarding.
The proxies now no longer know about setting up tunnels for forwarding application data. This is all moved into a 
seperate routing application. This application uses the (higher-level) 'VirtualSocketFactory', while the proxies use 
the (simpler/lower-level) 'DirectSocketFactory'. 

The advantage of this approach is that the routing application can now use all the features of the VirtualSocketFactory,
such as parallel streams or reverse connection setup. This was hard to do in the proxies themselves, since they operate
the DirectSocket level which doesn't offer these features...

When applications now register themselves at a proxy (by creating a ServiceLink), the fact that this client exists is 
gossiped between proxies. Clients can also advertise a certain service, by providing a name (String) and location 
(VirtualSocketAddress) to the service link. This advert will be gossiped along with the client info.

Routers use this information to find other routers. Similarly application can use this information to find their 
local routers, which they can then use to create connections to otherwhise unreachable machines.

Tested on 2 machine with a NAT box in between, which worked fine. 

Next steps: 

  - create module in VirtualSocketFactory which does this automagically
  - remove unused stuff from proxies
  - test on larger set of machines








git-svn-id: http://gforge.cs.vu.nl/svn/ibis/new-connect/trunk@3948 aaf88347-d911-0410-b711-e54d386773bb