Changed the proxy discovery such that the clients do an mcast to which the proxy replies (instead of the other way round). The advantage is that you can now easily run multiple clients on the same machine, because the port number on which they receive the discovery messages does no longer have to be fixed.
Added methods to SocketAddressSet which allow the user to check if two of 
them represent the same process or the same machine (NOTE: these are two 
different cases!)

Used these methods to check in the VirtualSocketFactory modules if a connection 
setup is requested to the same machine. This doesn't make any sense for the 
reverse/splice/routed modules (and didn't work either ;-).

This problem was trigger by an Ibis test that creates a Ibis + a nameserver from 
the same process. Sometimes the nameserver didn't start quickly enough which caused 
the client to try all sort (expensive) tricks to setup the connection. 

 

 


 

git-svn-id: http://gforge.cs.vu.nl/svn/ibis/new-connect/trunk@4336 aaf88347-d911-0410-b711-e54d386773bb