Fixed bug in VirtualSocketAddress(DataInputStream)
Slightly changed the sematics of the connect() in the virtual modules. Instead of
returning a 'finished' socket (i.e. one that has completed the connection setup), 
the modules now return a socket that is connected to the right machine, but still 
needs to perform the 'waitForAccept' operation. 

This approach makes it easier for the virtual socket layer to use the 'rest of the 
timeout' to finish the connection setup. It also allows the 'right' module to be 
retried immediately when a transient error occurs (such as an target overloaded 
exception). In these cases an exponential backoff mechanism is used.

 
 

git-svn-id: http://gforge.cs.vu.nl/svn/ibis/new-connect/trunk@5368 aaf88347-d911-0410-b711-e54d386773bb